Changes to the BLDC characterization code
I modified the ADC code so that the end of sampling is triggered
automatically after 1us. I also removed the code that read CH0. Adding
a 10uF decoupling capacitor eliminated the noise on the ADC lines.
However, when the PWM is running there is a 40kHz ripple voltage on the
UART TX line. I also added and clarified comments.